requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-bigquery==3.25.0
pydantic==2.8.2
python-dotenv==1.0.1  # Para carregar variáveis de ambiente em dev local